fix(vite-plugin): export the ./jsx-runtime subpath
The static JSX runtime documents `jsxImportSource:
"@ox-content/vite-plugin"`, which makes the compiler emit
`import { jsx } from "@ox-content/vite-plugin/jsx-runtime"`. That subpath was
never declared in `exports` and never built into `dist/`, so `tsc` could not
resolve the module and the emitted import failed at runtime. Consumers had to
keep a local `declare module` shim.

Adding the entry needs one more step than it looks. `src/jsx-runtime.ts` was
also imported by `src/index.ts` and `src/theme-renderer.ts`, so making it an
entry made the dts bundler take `jsx-runtime.d.mts` for the shared chunk (with
mangled export names) and push the real entry declarations to
`jsx-runtime2.d.mts`. The implementation therefore moves to `src/jsx-html.ts`,
leaving `./jsx-runtime` and `./jsx-dev-runtime` as thin entries that own their
own output names.

`./jsx-dev-runtime` ships too, since `jsx: "react-jsxdev"` imports it. Its
`jsxDEV` takes the transform's debug arguments and ignores them — static HTML
generation has nothing to do with them.

The `JSX` namespace now travels with the runtime exports instead of only being
declared globally, because `jsxImportSource` makes TypeScript look for the
element types in the runtime module itself. `key` moves onto `HTMLAttributes`
so `.map()` over a list of intrinsic elements type-checks; the automatic
runtime hoists it out of props and passes it as `jsx(type, props, key)`, where
it is ignored.

Verified end to end: `tsc` with both `react-jsx` and `react-jsxdev` emits
`import { jsx } from "@ox-content/vite-plugin/jsx-runtime"`, that import
resolves through ESM and CJS, and the compiled template renders.

Fixes #601
