+++ npm/vite-plugin-ox-content/vite.config.ts
@@ -6,7 +6,16 @@ export default defineConfig({
     ignorePatterns: ["dist/**"],
   },
   pack: definePackConfig({
-    entry: ["src/index.ts", "src/vitepress-cli.ts", "src/incremental-dom.ts"],
+    entry: [
+      "src/index.ts",
+      "src/vitepress-cli.ts",
+      "src/incremental-dom.ts",
+      // `jsxImportSource: "@ox-content/vite-plugin"` makes the JSX transform
+      // emit imports of ./jsx-runtime and ./jsx-dev-runtime, so both need to
+      // be real modules in dist/, not just symbols on the main entry.
+      "src/jsx-runtime.ts",
+      "src/jsx-dev-runtime.ts",
+    ],
     format: ["esm", "cjs"],
     target: "es2022",
     dts: true,
